turns out that removing the HttpSessionContextIntegrationFilter from dav and cmp requests introduced a bug where acegi security contexts were not released from server threads, so subsequent requests handled by those threads ignored any authentication credentials in the requests and used the prebound security contexts. bad stuff.
to counter this, we introduce the HttpRequestIntegrationFilter which
clears the security context at the end of a request. it is used by dav
and cmp. to facilitate this change, the acegi security config for the
three interfaces is broken into three separate xml files.

all of these changes exposed yet another subtle bug - because the ui
filter chain was being executed for the url path /*, it was being
executed on dav requests IN ADDITION TO the dav filter chain. fixing
this necessitated moving the web ui to the url path /console. now the
console filter chain only executes for /console/*, the dav filter
chain for /dav/*, etc.

as a side effect, this fixes bug 4045 and reduces the amount of url
rewriting that goes on to make the console's urls "friendly".


git-svn-id: http://svn.osafoundation.org/server/cosmo/branches/0.2@1102 5fd0b11a-f2f7-0310-948e-c8f80ebed640