bug 5095: synchronize access to the jcr template when creating home directories so that only one thread (and therefore jcr session) is modifying the root node at any given time. a better solution might be to use intermediary nodes between the root node and the homedir node, but even then there is still the possibility, however low, of concurrent modification of a parent node.
also, when deleting a user, ignore an exception signifying that the
home directory does not exist. if this is a broken account with a user
record but not a homedir node, the account will be deleted as
normal. if this is a regular old account that does not exist, then
attempting to remove the user record will throw an exception anyway.


git-svn-id: http://svn.osafoundation.org/server/cosmo/branches/0.2@1329 5fd0b11a-f2f7-0310-948e-c8f80ebed640