bug 5887: reimplement JcrUserDao.getUsers to directly iterate through the repository rather than using a jcr query.
tested against a 1000 user repository with a 64mb heap, this results
in approximately 100% performance gain (8 seconds to build the user
list as opposed to 16 with the jcr query).

compared to the 0.2 method of "select * from user" against hsqldb,
this operation is much slower. i think is due to a number of factors
but primarily the heavyweight stateful nature of the content
repository.

i think for now that we have to live with "list all users" being
slow. we have plans to modify the web console to only display subsets
of the full user list, and we should encourage administrators to use
this cmp operation rarely.

we can look in the future at changing the repository schema to store
user account info in a different subtree or even workspace than home
directory info. it's possible that queries/iteration would perform
better against user nodes that don't have content subtrees.



git-svn-id: http://svn.osafoundation.org/server/cosmo/branches/0.3@1980 5fd0b11a-f2f7-0310-948e-c8f80ebed640