fix for bug 8078 - End Date problems with Chandler/Cosmo
This one was tricky. Turns out that I was manipulating a date that was part of an EventStamp's calendar object 
for some date calcs that I needed. Problem is, the EventStamp object is kept in hibernate's 2nd level cache, 
along with its calendar. So if you screw up the calendar, you screw it up for all subsequent requests. 

Moral of the story- make a clone of the objects you need to mess around with, unless you will be persisting them back.

git-svn-id: http://svn.osafoundation.org/server/cosmo/branches/0.6.1@3701 5fd0b11a-f2f7-0310-948e-c8f80ebed640