Bug 7654: Add "Set" buttons to User List controls.
I had quite a bit of trouble getting our old style buttons to work here. First, using the buttons
in the most obvious way did not work because they are actually tables, and thus behave like divs.

Second, when I created a table to wrap all of these pieces, they simply didn't appear in IE. Since
we're moving to new "styled <input> buttons" for Preview I've just used non-styled buttons here for now.
Given that this is admin UI, I think this is a time-efficient short term solution.

Also, found that Safari was never displaying the PageNumberChooser control. Spent quite a bit of time wrestling with bug, the fix for which is explained in a comment in the code.


git-svn-id: http://svn.osafoundation.org/server/cosmo/branches/0.6.1@3821 5fd0b11a-f2f7-0310-948e-c8f80ebed640