Revert base64 encoding changes and change type attribute of content elements to "text/eim+json".
As discussed here:

http://lists.osafoundation.org/archives/cosmo-dev/200709.mbox/ajax/%3cC1815719-417A-4650-BA50-4A1BA2C5096C@osafoundation.org%3e

reverting the recently checked in code that base64 encoded the body of our Atom feed content elements. To maintain compatibility with the Atom processing model, also changing the type attribute of the content elements to "text/eim+json".

As noted in the thread, this change _will not_ be commited back to trunk.



git-svn-id: http://svn.osafoundation.org/server/cosmo/branches/0.7.2@5620 5fd0b11a-f2f7-0310-948e-c8f80ebed640