Merge r6363 to 0.11 branch. Original commit message was:
Basic admin ui tests, create a number of users with different characters in their names.

A good amount of refactoring/cleanup was needed to make this easily testable.

Tests are currently untested, as I couldn't figure out how to get them to work. Will chat with QA guys on Monday to rectify this.


git-svn-id: http://svn.osafoundation.org/server/cosmo/branches/0.11@6364 5fd0b11a-f2f7-0310-948e-c8f80ebed640